requirements.txt
ultralytics
easyocr
appdirs==1.4.4
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
bs4==0.0.2
CacheControl==0.14.0
cachetools==5.4.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cryptography==43.0.0
cycler==0.12.1
exceptiongroup==1.2.1
filelock==3.15.4
firebase-admin==6.5.0
Flask==3.0.3
Flask-Cors==4.0.1
fonttools==4.53.1
fsspec==2024.6.1
fuzzywuzzy==0.18.0
google-api-core==2.19.1
google-api-python-client==2.138.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.17.0
google-cloud-storage==2.18.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
grpcio==1.65.1
grpcio-status==1.62.2
h11==0.14.0
httplib2==0.22.0
idna==3.7
imageio==2.34.2
importlib_metadata==8.0.0
importlib_resources==6.4.0
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.5
lazy_loader==0.4
Levenshtein==0.25.1
livestreamer==1.12.2
lxml==5.2.2
MarkupSafe==2.1.5
matplotlib==3.9.1
MouseInfo==0.1.3
mpmath==1.3.0
msgpack==1.0.8
networkx==3.2.1
ninja==1.11.1.1
numpy==2.0.0
obs-websocket-py==1.0
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.2
pillow==10.4.0
proto-plus==1.24.0
protobuf==4.25.4
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyAutoGUI==0.9.54
pyclipper==1.3.0.post5
pycountry==24.6.1
pycparser==2.22
pycryptodome==3.20.0
pyee==11.1.0
PyGetWindow==0.0.9
PyJWT==2.8.0
PyMsgBox==1.0.9
pyparsing==3.1.2
pyperclip==1.9.0
pyppeteer==2.0.0
PyRect==0.2.0
PyScreeze==0.1.30
PySocks==1.7.1
pytesseract==0.3.10
python-bidi==0.5.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-Levenshtein==0.25.1
pytube==15.0.0
pytweening==1.2.0
pytz==2024.1
PyYAML==6.0.1
rapidfuzz==3.9.4
requests==2.32.3
rsa==4.9
rubicon-objc==0.4.9
schedule==1.2.2
scikit-image==0.24.0
scipy==1.13.1
seaborn==0.13.2
selenium==4.22.0
shapely==2.0.5
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
streamlink==6.8.3
sympy==1.13.1
tifffile==2024.7.21
torch==2.3.1
torchvision==0.18.1
tqdm==4.66.4
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==1.26.19
waitress==3.0.0
webdriver-manager==4.0.1
websocket-client==1.8.0
websockets==10.4
Werkzeug==3.0.3
wsproto==1.2.0
zipp==3.19.2
redis>=4.0.0
kombu>=5.2.0
